Moved User to Models
